global: use uri/uri-reference formats, not url
* INCOMPATIBLE This replaces the `url` `format` everywhere by `uri` (for
absolute URLs) or `uri-reference` (if the URL could be relative).  These
`format`s are part of the JSON schema spec (the second one starting from
Draft 6), whereas `url` is not. In order to validate these formats, this
package now depends on `rfc3987`, and adds a custom validator for
`uri-reference`, as we are currently on Draft 4 (closes #274).

Note that the `json-schema-faker` currently generates relative URIs for
`uri` (which is incompatible with Draft 6), so the fixtures had to be
amended by hand.

Signed-off-by: Micha Moskovic <michamos@gmail.com>